add lightning effect, add alignment level charges for each perk where necessary

+++ src/main/java/hellfirepvp/astralsorcery/common/data/research/PlayerProgress.java
@@ -212,7 +212,7 @@ public class PlayerProgress {
         return ConstellationPerkLevelManager.getAlignmentLevel(this);
     }
 
-    protected void modifyCharge(int charge) {
+    protected void modifyCharge(double charge) {
         this.alignmentCharge = Math.max(0, this.alignmentCharge + charge);
     }
